Start consensus, service handle request

+++ HDS/Blockchain/src/main/java/pt/tecnico/ulisboa/Account.java
@@ -5,20 +5,24 @@ import java.security.*;
 public class Account {
     private int balance;
     private final PublicKey publicKey;
+
     public Account(PublicKey publicKey) {
         this.publicKey = publicKey;
         this.balance = 20;
     }
+
     public int check_balance(){
         return this.balance;
     }
-    public void removeBalance(double removeBalance){
+
+    public void removeBalance(int removeBalance){
         this.balance -= removeBalance;
     }
 
-    public void addBalance(double addedBalance){
+    public void addBalance(int addedBalance){
         this.balance += addedBalance;
     }
+
     public String getPublicKey(){
         return this.publicKey.toString();
     }
